CLI: add the `scrape` command to scrape CIFs and store in AiiDA
This command will scrape the contents of the AMCSD mineral index to get
a list of available minerals. For each mineral, its dedicated page will
then be scraped for anchors that allow downloading its structure in CIF
format. These are then downloaded and wrapped in `SinglefileData` nodes
that are added to a `Group`. The source metadata and md5 of the
retrieved CIF files are added as attributes to the `SinglefileData`.

The `requests-cache` library is used to cache the requests such that the
server is not overloaded on subsequent invocations of the command.