CLI: add the `parse` command to parse scraped and imported CIF files
The command will loop over the `SinglefileData` nodes of the group
provided by the `-F/--group-cif-file` options. Each node is then passed
to the `cif_file_to_cif_data` calcfunction.

The calcfunction attempts to construct a `CifData` node from the
contents of the CIF file stored in the `SinglefileData` node. The CIF
files provided by AMCSD are known to have innocent syntax errors in the
publication information which will cause the parsing of the CIF to fail.
If such an error occurs a list of regexes is applied that correct these
known mistakes and parsing is tried again.

Finally, the created `CifData` nodes are stored in the group that was
specified with the `-D/--group-cif-data` option.