CLI: add the `-v/--verbosity` option to the root command
This option will allow to control the verbosity of all commands. It is
defined by `aiida-core` and by using this option, it allows us directly
to use the `aiida.cmdline.utils.echo` module that provides useful
shortcuts to echo messages of different type and styling.

Note that these methods only work if the logging of AiiDA's command line
logger is properly configured. This is done automatically by using the
predefined `aiida.cmdline.params.options.VERBOSITY` option.