CLI: deduplicate CIF files based on md5 in `scrape`
Each mineral page is searched for links that allow downloading a CIF
file. The current implementation would stop as soon as one CIF file was
successfully downloaded. However, certain pages contain multiple links
that may have more than one valid CIF to be downloaded, so the break
after the first downloaded CIF is removed.

That being said, in some cases the CIF files may actually be identical
and so to prevent loading duplicates in the database, the md5 of the
file is computed to check if it is already contained in the database or
the CIF files that were just loaded during the command execution.